# server/docs/swagger.yaml
definitions:
  app.Category:
    properties:
      ID:
        description: 主键ID
        type: integer
      categoryID:
        description: categoryID
        type: string
      createdAt:
        description: 创建时间
        type: string
      description:
        description: description
        type: string
      name:
        description: name
        type: string
      updatedAt:
        description: 更新时间
        type: string
    type: object
  app.Product:
    properties:
      ID:
        description: 主键ID
        type: integer
      categoryID:
        description: categoryID
        type: string
      categoryName:
        description: 类别名称
        type: string
      createdAt:
        description: 创建时间
        type: string
      description:
        description: description
        type: string
      imgs:
        description: imgs
        items:
          type: object
        type: array
      isDelete:
        description: isDelete
        type: integer
      name:
        description: name
        type: string
      price:
        description: price
        type: integer
      productID:
        description: productID
        type: string
      stock:
        description: stock
        type: integer
      supplierID:
        description: supplierID
        type: string
      updatedAt:
        description: 更新时间
        type: string
    type: object
...
  response.PageResult:
    properties:
      list: {}
      page:
        type: integer
      pageSize:
        type: integer
      total:
        type: integer
    type: object
...
  response.Response:
    properties:
      code:
        type: integer
      data: {}
      msg:
        type: string
    type: object
...
info:
  contact: {}
  description: 使用gin+vue进行极速开发的全栈开发基础平台
  title: Gin-Vue-Admin Swagger API接口文档
  version: v2.8.0
paths:
...
  /category/createCategory:
    post:
      consumes:
      - application/json
      parameters:
      - description: 创建category
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/app.Category'
      produces:
      - application/json
      responses:
        "200":
          description: 创建成功
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                msg:
                  type: string
              type: object
      security:
      - ApiKeyAuth: []
      summary: 创建category
      tags:
      - Category
  /category/deleteCategory:
    delete:
      consumes:
      - application/json
      parameters:
      - description: 删除category
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/app.Category'
      produces:
      - application/json
      responses:
        "200":
          description: 删除成功
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                msg:
                  type: string
              type: object
      security:
      - ApiKeyAuth: []
      summary: 删除category
      tags:
      - Category
  /category/deleteCategoryByIds:
    delete:
      consumes:
      - application/json
      produces:
      - application/json
      responses:
        "200":
          description: 批量删除成功
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                msg:
                  type: string
              type: object
      security:
      - ApiKeyAuth: []
      summary: 批量删除category
      tags:
      - Category
  /category/findCategory:
    get:
      consumes:
      - application/json
      parameters:
      - description: 用id查询category
        in: query
        name: ID
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 查询成功
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/app.Category'
                msg:
                  type: string
              type: object
      security:
      - ApiKeyAuth: []
      summary: 用id查询category
      tags:
      - Category
  /category/getCategoryList:
    get:
      consumes:
      - application/json
      parameters:
      - in: query
        name: categoryID
        type: string
      - in: query
        name: endCreatedAt
        type: string
      - description: 关键字
        in: query
        name: keyword
        type: string
      - in: query
        name: name
        type: string
      - description: 页码
        in: query
        name: page
        type: integer
      - description: 每页大小
        in: query
        name: pageSize
        type: integer
      - in: query
        name: startCreatedAt
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 获取成功
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/response.PageResult'
                msg:
                  type: string
              type: object
      security:
      - ApiKeyAuth: []
      summary: 分页获取category列表
      tags:
      - Category
...
  /category/updateCategory:
    put:
      consumes:
      - application/json
      parameters:
      - description: 更新category
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/app.Category'
      produces:
      - application/json
      responses:
        "200":
          description: 更新成功
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                msg:
                  type: string
              type: object
      security:
      - ApiKeyAuth: []
      summary: 更新category
      tags:
      - Category
...
  /product/createProduct:
    post:
      consumes:
      - application/json
      parameters:
      - description: 创建商品
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/app.Product'
      produces:
      - application/json
      responses:
        "200":
          description: 创建成功
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                msg:
                  type: string
              type: object
      security:
      - ApiKeyAuth: []
      summary: 创建商品
      tags:
      - Product
  /product/deleteProduct:
    delete:
      consumes:
      - application/json
      parameters:
      - description: 删除商品
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/app.Product'
      produces:
      - application/json
      responses:
        "200":
          description: 删除成功
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                msg:
                  type: string
              type: object
      security:
      - ApiKeyAuth: []
      summary: 删除商品
      tags:
      - Product
  /product/deleteProductByIds:
    delete:
      consumes:
      - application/json
      produces:
      - application/json
      responses:
        "200":
          description: 批量删除成功
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                msg:
                  type: string
              type: object
      security:
      - ApiKeyAuth: []
      summary: 批量删除商品
      tags:
      - Product
  /product/findProduct:
    get:
      consumes:
      - application/json
      parameters:
      - description: 用id查询商品
        in: query
        name: ID
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 查询成功
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/app.Product'
                msg:
                  type: string
              type: object
      security:
      - ApiKeyAuth: []
      summary: 用id查询商品
      tags:
      - Product
  /product/getProductList:
    get:
      consumes:
      - application/json
      parameters:
      - in: query
        name: categoryID
        type: string
      - in: query
        name: categoryName
        type: string
      - in: query
        name: endCreatedAt
        type: string
      - description: 关键字
        in: query
        name: keyword
        type: string
      - in: query
        name: name
        type: string
      - description: 页码
        in: query
        name: page
        type: integer
      - description: 每页大小
        in: query
        name: pageSize
        type: integer
      - in: query
        name: price
        type: integer
      - in: query
        name: productID
        type: string
      - in: query
        name: startCreatedAt
        type: string
      - in: query
        name: supplierID
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 获取成功
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/response.PageResult'
                msg:
                  type: string
              type: object
      security:
      - ApiKeyAuth: []
      summary: 分页获取商品列表
      tags:
      - Product
...
  /product/updateProduct:
    put:
      consumes:
      - application/json
      parameters:
      - description: 更新商品
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/app.Product'
      produces:
      - application/json
      responses:
        "200":
          description: 更新成功
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                msg:
                  type: string
              type: object
      security:
      - ApiKeyAuth: []
      summary: 更新商品
      tags:
      - Product
...
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: x-token
    type: apiKey
swagger: "2.0"
...